UI edits made (final version)

+++ MovieWatchlistTracker/src/MovieWatchlistTracker.Web/Services/Interfaces/IRatingService.cs
@@ -4,7 +4,7 @@ namespace MovieWatchlistTracker.Web.Services.Interfaces;
 
 public interface IRatingService
 {
-    Task<Rating> CreateOrUpdateAsync(string userId, int movieId, int score);
+    Task<Rating> CreateOrUpdateAsync(string userId, int movieId, double score);
     Task<bool> DeleteAsync(int ratingId, string userId);
     Task<bool> UserOwnsRatingAsync(int ratingId, string userId);
 }
